refactor(mongo-user-model): rename 'user-model$' to 'mongo-user-model'
